refactor(cli): move the builtin command registry into an importable module
bin.ts defined every command and the registry in the same module that discovers plugin commands from process.cwd() and calls runCli() at top level, so nothing could read the registry without running the CLI. The agent-catalog audit (RFC 0011 §2) needs to assert that a published skill names only commands and flags the CLI registers.

Mechanical split: everything through `builtinSubCommands` moves to commands.ts and is exported; bin.ts keeps the shebang, plugin discovery, readOwnVersion, main, and runCli. No definition changed. --help still lists 67 commands, --version still works, and importing commands.ts alone builds the registry in ~100ms with no side effects.
